+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Update the `org.realityforge.guiceyloops` artifact to version `0.113`.
 * Update the java version required to `17`.
+* Remove several typos in `dependencies.yml`
 
 ### [v0.17](https://github.com/realityforge/bazel-depgen/tree/v0.17) (2022-05-24) · [Full Changelog](https://github.com/spritz/spritz/compare/v0.16...v0.17)